+++ spec/models/descriptor/qualitative_spec.rb
@@ -14,4 +14,13 @@ RSpec.describe Descriptor::Qualitative, type: :model, group: [:descriptor, :obse
     expect(descriptor.character_states.count).to eq(2)
   end
 
+  specify 'destroys states if unused' do
+    descriptor.name = 'Head color'
+    descriptor.character_states_attributes = [ {name: 'Blue', label: '0'}, {name: 'Red', label: '1'} ] 
+    descriptor.save!
+    descriptor.destroy
+    expect(CharacterState.all.count).to eq(0) 
+  end
+
+
 end